requirements.txt
pytest==8.2.0
httpx==0.27.0
uvicorn[standard]==0.29.0
fastapi==0.110.2
tweepy==4.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
git+https://github.com/joeljunior95/owmpy.git